@@ -1,9 +1,10 @@
 {
     "name": "zeebe-node-affinity",
     "version": "0.20.0",
-    "description": "Zeebe node client and affinity service to communicate workflow outcome to its initiator via a web socket subscription",
+    "description": "Zeebe node client and affinity service to communicate process outcome to its initiator via a web socket subscription",
     "main": "dst/index.js",
     "scripts": {
+        "lint": "npx eslint . --ext .ts --fix",
         "test": "echo \"Error: no test specified\" && exit 1",
         "build": "tsc"
     },
@@ -26,9 +27,19 @@
         "dayjs": "^1.8.15",
         "promise-retry": "^1.1.1",
         "redis": "^3.0.2",
-        "typescript": "^4.3.5",
         "uuid": "^8.3.2",
         "ws": "^7.1.0",
-        "zeebe-node": "^0.26.0"
+        "zeebe-node": "^1.3.4"
+    },
+    "devDependencies": {
+        "@types/node": "^16.10.3",
+        "@typescript-eslint/eslint-plugin": "^4.33.0",
+        "eslint": "^7.32.0",
+        "eslint-plugin-import": "^2.24.2",
+        "eslint-plugin-prettier": "^4.0.0",
+        "eslint-plugin-promise": "^5.1.0",
+        "prettier": "^2.4.1",
+        "ts-node": "^10.2.1",
+        "typescript": "^4.4.3"
     }
 }
