@@ -1,9 +1,9 @@
 /// <reference types="node" />
-import WebSocket from "ws";
-import { ZBClient } from "zeebe-node";
-import { KeyedObject } from "zeebe-node/dist/lib/interfaces";
+import WebSocket from 'ws';
+import { ZBClient } from 'zeebe-node';
+import { KeyedObject } from 'zeebe-node/dist/lib/interfaces';
 import { ZBClientOptions } from 'zeebe-node/dist/lib/interfaces-published-contract';
-import { WorkflowOutcome } from "./WebSocketAPI";
+import { ProcessOutcome } from './WebSocketAPI';
 interface ZBAffinityClientOptions extends ZBClientOptions {
     affinityServiceUrl: string;
     affinityTimeout: number;
@@ -13,18 +13,18 @@ export declare class ZBAffinityClient extends ZBClient {
     affinityService: WebSocket;
     ws: any;
     affinityCallbacks: {
-        [workflowInstanceKey: string]: (workflowOutcome: WorkflowOutcome) => void;
+        [processInstanceKey: string]: (processOutcome: ProcessOutcome) => void;
     };
     affinityTimeout: number;
     pingTimeout: NodeJS.Timer;
     constructor(gatewayAddress: string, options: ZBAffinityClientOptions);
     createAffinityWorker(taskType: string): Promise<void>;
-    createWorkflowInstanceWithAffinity<Variables = KeyedObject>({ bpmnProcessId, variables, version, cb }: {
+    createProcessInstanceWithAffinity<Variables = KeyedObject>({ bpmnProcessId, variables, version, cb, }: {
         bpmnProcessId: string;
         variables: Variables;
         version?: number;
-        cb: (workflowOutcome: WorkflowOutcome) => void;
-    }): Promise<import("zeebe-node").CreateWorkflowInstanceResponse>;
+        cb: (processOutcome: ProcessOutcome) => void;
+    }): Promise<import("zeebe-node/dist/lib/interfaces-grpc-1.0").CreateProcessInstanceResponse>;
     waitForAffinity(): Promise<void>;
     private throwNoConnection;
     private createAffinityService;
