@@ -1,67 +1,67 @@
-import WebSocket from "ws";
+import WebSocket from 'ws';
 
 export enum AffinityAPIMessageType {
-  WORKFLOW_OUTCOME = "WORKFLOW_OUTCOME",
-  REGISTER_WORKER = "REGISTER_WORKER",
-  REGISTER_CLIENT = "REGISTER_CLIENT"
+    PROCESS_OUTCOME = 'PROCESS_OUTCOME',
+    REGISTER_WORKER = 'REGISTER_WORKER',
+    REGISTER_CLIENT = 'REGISTER_CLIENT',
 }
 
-export interface WorkflowOutcomeMessage {
-  type: AffinityAPIMessageType.WORKFLOW_OUTCOME;
-  workflowInstanceKey: string;
-  variables: { [key: string]: string | number };
+export interface ProcessOutcomeMessage {
+    type: AffinityAPIMessageType.PROCESS_OUTCOME;
+    processInstanceKey: string;
+    variables: { [key: string]: string | number };
 }
 
 export interface RegisterClientMessage {
-  type: AffinityAPIMessageType.REGISTER_CLIENT;
+    type: AffinityAPIMessageType.REGISTER_CLIENT;
 }
 
 export interface RegisterWorkerMessage {
-  type: AffinityAPIMessageType.REGISTER_WORKER;
+    type: AffinityAPIMessageType.REGISTER_WORKER;
 }
 
-export interface WorkflowOutcome {
-  workflowInstanceKey: string;
-  variables: { [key: string]: string | number };
+export interface ProcessOutcome {
+    processInstanceKey: string;
+    variables: { [key: string]: string | number };
 }
 
 export function registerWorker(ws: WebSocket) {
-  ws.send(JSON.stringify({ type: AffinityAPIMessageType.REGISTER_WORKER }));
+    ws.send(JSON.stringify({ type: AffinityAPIMessageType.REGISTER_WORKER }));
 }
 
 export function registerClient(ws: WebSocket) {
-  ws.send(JSON.stringify({ type: AffinityAPIMessageType.REGISTER_CLIENT }));
+    ws.send(JSON.stringify({ type: AffinityAPIMessageType.REGISTER_CLIENT }));
 }
 
-export function broadcastWorkflowOutcome(
-  clients: { [uuid: string]: WebSocket },
-  workflowOutcome: WorkflowOutcome
+export function broadcastProcessOutcome(
+    clients: { [uuid: string]: WebSocket },
+    processOutcome: ProcessOutcome,
 ) {
-  const message: WorkflowOutcomeMessage = {
-    type: AffinityAPIMessageType.WORKFLOW_OUTCOME,
-    ...workflowOutcome
-  };
-  Object.values(clients).forEach(client => {
-    if (client.readyState === WebSocket.OPEN) {
-      client.send(JSON.stringify(message));
-    }
-  });
+    const message: ProcessOutcomeMessage = {
+        type: AffinityAPIMessageType.PROCESS_OUTCOME,
+        ...processOutcome,
+    };
+    Object.values(clients).forEach((client) => {
+        if (client.readyState === WebSocket.OPEN) {
+            client.send(JSON.stringify(message));
+        }
+    });
 }
 
-export function demarshalWorkflowOutcome(data): WorkflowOutcome | undefined {
-  const message = JSON.parse(data.toString());
-  return (message.type = AffinityAPIMessageType.WORKFLOW_OUTCOME
-    ? { ...message, type: undefined }
-    : undefined);
+export function demarshalProcessOutcome(data): ProcessOutcome | undefined {
+    const message = JSON.parse(data.toString());
+    return (message.type = AffinityAPIMessageType.PROCESS_OUTCOME
+        ? { ...message, type: undefined }
+        : undefined);
 }
 
-export function publishWorkflowOutcomeToAffinityService(
-  workflowOutcome: WorkflowOutcome,
-  ws
+export function publishProcessOutcomeToAffinityService(
+    processOutcome: ProcessOutcome,
+    ws,
 ) {
-  const workflowOutcomeMessage: WorkflowOutcomeMessage = {
-    type: AffinityAPIMessageType.WORKFLOW_OUTCOME,
-    ...workflowOutcome
-  };
-  ws.send(JSON.stringify(workflowOutcomeMessage));
+    const processOutcomeMessage: ProcessOutcomeMessage = {
+        type: AffinityAPIMessageType.PROCESS_OUTCOME,
+        ...processOutcome,
+    };
+    ws.send(JSON.stringify(processOutcomeMessage));
 }
